+++ service-member/module-api/src/main/java/com/github/msafriends/moduleapi/service/MemberCouponService.java
@@ -5,14 +5,15 @@ import com.github.msafriends.moduleapi.dto.request.member.MemberCouponRequest;
 import com.github.msafriends.moduleapi.dto.request.member.MemberCouponUseRequest;
 import com.github.msafriends.moduleapi.dto.response.ListResponse;
 import com.github.msafriends.moduleapi.dto.response.coupon.MemberCouponResponse;
-import com.github.msafriends.modulecommon.exception.member.coupon.CouponAlreadyIssuedException;
-import com.github.msafriends.modulecommon.exception.member.coupon.CouponExceedLimitedQuantityException;
-import com.github.msafriends.modulecommon.exception.member.coupon.CouponNotExistException;
-import com.github.msafriends.modulecommon.exception.member.coupon.CouponOutOfStockException;
+
 import com.github.msafriends.modulecore.domain.coupon.Coupon;
 import com.github.msafriends.modulecore.domain.coupon.CouponGenerateType;
 import com.github.msafriends.modulecore.domain.coupon.MemberCoupon;
 import com.github.msafriends.modulecore.domain.member.Member;
+import com.github.msafriends.modulecore.exception.member.coupon.CouponAlreadyIssuedException;
+import com.github.msafriends.modulecore.exception.member.coupon.CouponExceedLimitedQuantityException;
+import com.github.msafriends.modulecore.exception.member.coupon.CouponNotExistException;
+import com.github.msafriends.modulecore.exception.member.coupon.CouponOutOfStockException;
 import com.github.msafriends.modulecore.repository.coupon.CouponRepository;
 import com.github.msafriends.modulecore.repository.coupon.MemberCouponRepository;
 import com.github.msafriends.modulecore.repository.member.MemberRepository;
