+++ service-product/module-core/src/main/java/com/github/msafriends/serviceproduct/modulecore/repository/ProductRepository.java
@@ -1,16 +1,14 @@
 package com.github.msafriends.serviceproduct.modulecore.repository;
 
 
-import com.github.msafriends.modulecommon.exception.EntityNotFoundException;
-import com.github.msafriends.modulecommon.exception.ErrorCode;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Lock;
 import org.springframework.data.jpa.repository.Query;
 
-import com.github.msafriends.modulecommon.exception.EntityNotFoundException;
-import com.github.msafriends.modulecommon.exception.ErrorCode;
 import com.github.msafriends.serviceproduct.modulecore.domain.product.Product;
+import com.github.msafriends.serviceproduct.modulecore.exception.EntityNotFoundException;
+import com.github.msafriends.serviceproduct.modulecore.exception.ErrorCode;
 
 import jakarta.persistence.LockModeType;
