+++ service-product/module-core/src/main/java/com/github/msafriends/serviceproduct/modulecore/repository/ProductReviewRepository.java
@@ -5,9 +5,10 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import com.github.msafriends.modulecommon.exception.EntityNotFoundException;
-import com.github.msafriends.modulecommon.exception.ErrorCode;
+
 import com.github.msafriends.serviceproduct.modulecore.domain.review.ProductReview;
+import com.github.msafriends.serviceproduct.modulecore.exception.EntityNotFoundException;
+import com.github.msafriends.serviceproduct.modulecore.exception.ErrorCode;
 
 public interface ProductReviewRepository extends JpaRepository<ProductReview, Long> {
